Migrates defi module to pinia
