+++ frontend/app/src/components/defi/Borrowing.vue
@@ -7,7 +7,7 @@
       <v-col>
         <refresh-header
           :title="$t('borrowing.header')"
-          :loading="anyRefreshing"
+          :loading="refreshing"
           @refresh="refresh()"
         >
           <template #actions>
@@ -25,7 +25,7 @@
                 {{ $t('borrowing.total_collateral_locked') }}
               </template>
               <amount-display
-                :value="loanSummary(selectedProtocols).totalCollateralUsd"
+                :value="summary.totalCollateralUsd"
                 show-currency="symbol"
                 fiat-currency="USD"
               />
@@ -37,7 +37,7 @@
                 {{ $t('borrowing.total_outstanding_debt') }}
               </template>
               <amount-display
-                :value="loanSummary(selectedProtocols).totalDebt"
+                :value="summary.totalDebt"
                 show-currency="symbol"
                 fiat-currency="USD"
               />
@@ -55,8 +55,10 @@
               class="borrowing__vault-selection"
               :label="$t('borrowing.select_loan')"
               chips
+              dense
+              outlined
               item-key="identifier"
-              :items="loans(selectedProtocols)"
+              :items="loans"
               item-text="identifier"
               hide-details
               clearable
@@ -77,7 +79,7 @@
         <defi-protocol-selector v-model="protocol" liabilities />
       </v-col>
     </v-row>
-    <loan-info v-if="selection" :loan="loan(selection)" />
+    <loan-info v-if="selection" :loan="loan" />
     <full-size-content v-else>
       <v-row align="center" justify="center">
         <v-col class="text-h6">{{ $t('liabilities.no_selection') }}</v-col>
@@ -88,9 +90,14 @@
 
 <script lang="ts">
 import { DefiProtocol } from '@rotki/common/lib/blockchain';
-import { PropType } from 'vue';
-import { Component, Mixins, Prop } from 'vue-property-decorator';
-import { mapActions, mapGetters } from 'vuex';
+import {
+  computed,
+  defineComponent,
+  onMounted,
+  PropType,
+  ref
+} from '@vue/composition-api';
+import { get, set } from '@vueuse/core';
 import FullSizeContent from '@/components/common/FullSizeContent.vue';
 import ActiveModules from '@/components/defi/ActiveModules.vue';
 import DefiSelectorItem from '@/components/defi/DefiSelectorItem.vue';
@@ -101,23 +108,13 @@ import StatCardWide from '@/components/display/StatCardWide.vue';
 import DefiProtocolSelector from '@/components/helper/DefiProtocolSelector.vue';
 import ProgressScreen from '@/components/helper/ProgressScreen.vue';
 import RefreshHeader from '@/components/helper/RefreshHeader.vue';
-import StatusMixin from '@/mixins/status-mixin';
+import { setupStatusChecking, useRoute } from '@/composables/common';
 import { Section } from '@/store/const';
-import {
-  AaveLoan,
-  DefiLoan,
-  LoanSummary,
-  MakerDAOVaultModel
-} from '@/store/defi/types';
+import { useDefiSupportedProtocolsStore } from '@/store/defi/protocols';
 import { Module } from '@/types/modules';
 
-@Component({
-  computed: {
-    ...mapGetters('defi', ['loan', 'loans', 'loanSummary'])
-  },
-  methods: {
-    ...mapActions('defi', ['fetchBorrowing'])
-  },
+export default defineComponent({
+  name: 'Borrowing',
   components: {
     FullSizeContent,
     ActiveModules,
@@ -129,40 +126,71 @@ import { Module } from '@/types/modules';
     StatCardWide,
     LoanInfo,
     ProgressScreen
+  },
+  props: {
+    modules: { required: true, type: Array as PropType<Module[]> }
+  },
+  setup() {
+    const selection = ref<string>();
+    const protocol = ref<DefiProtocol | null>(null);
+    const store = useDefiSupportedProtocolsStore();
+    const route = useRoute();
+
+    const { shouldShowLoadingScreen, isSectionRefreshing } =
+      setupStatusChecking();
+
+    const selectedProtocols = computed(() => {
+      const selected = get(protocol);
+      return selected ? [selected] : [];
+    });
+
+    const loan = computed(() => get(store.loan(get(selection))));
+
+    const loans = computed(() => {
+      const protocols = get(selectedProtocols);
+      return get(store.loans(protocols));
+    });
+
+    const summary = computed(() => {
+      const protocols = get(selectedProtocols);
+      return get(store.loanSummary(protocols));
+    });
+
+    const refreshing = computed(() => {
+      return (
+        get(isSectionRefreshing(Section.DEFI_BORROWING)) ||
+        get(isSectionRefreshing(Section.DEFI_BORROWING_HISTORY))
+      );
+    });
+
+    const refresh = async () => {
+      await store.fetchBorrowing(true);
+    };
+
+    onMounted(async () => {
+      const currentRoute = get(route);
+      const queryElement = currentRoute.query['protocol'];
+      const protocols = Object.values(DefiProtocol);
+      const protocolIndex = protocols.findIndex(
+        protocol => protocol === queryElement
+      );
+      if (protocolIndex >= 0) {
+        set(protocol, protocols[protocolIndex]);
+      }
+      await store.fetchBorrowing(false);
+    });
+
+    return {
+      selection,
+      protocol,
+      selectedProtocols,
+      loan,
+      loans,
+      summary,
+      loading: shouldShowLoadingScreen(Section.DEFI_BORROWING),
+      refreshing,
+      refresh
+    };
   }
-})
-export default class Liabilities extends Mixins(StatusMixin) {
-  selection?: string = '';
-  loan!: (identifier?: string) => MakerDAOVaultModel | AaveLoan | null;
-  loans!: (protocol: DefiProtocol[]) => DefiLoan[];
-  loanSummary!: (protocol: DefiProtocol[]) => LoanSummary;
-  fetchBorrowing!: (refreshing: boolean) => Promise<void>;
-  protocol: DefiProtocol | null = null;
-
-  section = Section.DEFI_BORROWING;
-  secondSection = Section.DEFI_BORROWING_HISTORY;
-
-  @Prop({ required: true, type: Array as PropType<Module[]> })
-  modules!: Module[];
-
-  get selectedProtocols(): DefiProtocol[] {
-    return this.protocol ? [this.protocol] : [];
-  }
-
-  async created() {
-    const queryElement = this.$route.query['protocol'];
-    const protocols = Object.values(DefiProtocol);
-    const protocolIndex = protocols.findIndex(
-      protocol => protocol === queryElement
-    );
-    if (protocolIndex >= 0) {
-      this.protocol = protocols[protocolIndex];
-    }
-    await this.fetchBorrowing(false);
-  }
-
-  async refresh() {
-    await this.fetchBorrowing(true);
-  }
-}
+});
 </script>
