+++ frontend/app/src/store/defi/const.ts
@@ -1,8 +1,6 @@
 import { balanceKeys } from '@/services/consts';
 import { OverviewDefiProtocol } from '@/store/defi/types';
 
-export const ACTION_PURGE_PROTOCOL = 'purgeProtocol' as const;
-
 const FINNEXUS = 'FinNexus';
 const PICKLE_FINANCE = 'Pickle Finance';
 const PICKLE_FINANCE_FARMS = 'Pickle Finance • Farms';
