--- frontend/app/src/store/defi/state.ts
@@ -1,8 +1,0 @@
-import { DefiState } from '@/store/defi/types';
-
-export const defaultState = (): DefiState => ({
-  allProtocols: {},
-  airdrops: {}
-});
-
-export const state: DefiState = defaultState();
