+++ frontend/app/src/views/defi/DecentralizedBorrowing.vue
@@ -1,31 +1,21 @@
 <template>
   <module-not-active v-if="!anyModuleEnabled" :modules="modules" />
-  <liabilities v-else :modules="modules" />
+  <borrowing v-else :modules="modules" />
 </template>
 
-<script lang="ts">
-import { defineComponent } from '@vue/composition-api';
-import Liabilities from '@/components/defi/Liabilities.vue';
+<script setup lang="ts">
+import Borrowing from '@/components/defi/Borrowing.vue';
 import ModuleNotActive from '@/components/defi/ModuleNotActive.vue';
 import { useModules } from '@/composables/session';
 import { Module } from '@/types/modules';
 
-export default defineComponent({
-  components: { ModuleNotActive, Liabilities },
-  setup() {
-    const modules: Module[] = [
-      Module.AAVE,
-      Module.COMPOUND,
-      Module.MAKERDAO_VAULTS,
-      Module.LIQUITY
-    ];
+const modules: Module[] = [
+  Module.AAVE,
+  Module.COMPOUND,
+  Module.MAKERDAO_VAULTS,
+  Module.LIQUITY
+];
 
-    const { isAnyModuleEnabled } = useModules();
-
-    return {
-      modules,
-      anyModuleEnabled: isAnyModuleEnabled(modules)
-    };
-  }
-});
+const { isAnyModuleEnabled } = useModules();
+const anyModuleEnabled = isAnyModuleEnabled(modules);
 </script>
